+++ src/components/Cards/Cards.module.css
@@ -28,14 +28,11 @@
 .header {
   display: flex;
   justify-content: space-between;
-  align-items: end;
   margin-bottom: 35px;
 }
 
 .timer {
   display: flex;
-  align-items: end;
-
   color: #fff;
   font-family: StratosSkyeng;
   font-size: 64px;
